Start working on implementing support for planar PCS soft robot datasets

+++ src/dataset_collection.py
@@ -14,7 +14,7 @@ def collect_dataset(
     rendering_fn: Callable,
     rng: random.KeyArray,
     num_simulations: int,
-    horizon: Array,
+    horizon_dim: int,
     dt: Array,
     state_init_min: Array,
     state_init_max: Array,
@@ -22,6 +22,8 @@ def collect_dataset(
     solver: AbstractSolver = Dopri5(),
     sim_dt: Optional[Array] = None,
     system_params: Dict[str, Array] = None,
+    do_yield: bool = True,
+    save_raw_data: bool = False,
 ):
     """
     Collect a simulated dataset for a given ODE system. The initial state is uniformly sampled from the given bounds.
@@ -32,7 +34,7 @@ def collect_dataset(
             rendering_fn(q) -> img
         rng: PRNG key for random number generation.
         num_simulations: Number of simulations to run.
-        horizon: Duration of each simulation [s].
+        horizon_dim: Number of samples in each trajectory.
         dt: Time step used for samples [s].
         state_init_min: Array with minimal values for the initial state of the simulation.
         state_init_max: Array with maximal values for the initial state of the simulation.
@@ -40,12 +42,14 @@ def collect_dataset(
         solver: Diffrax solver to use for the simulation.
         sim_dt: Time step used for simulation [s].
         system_params: Dictionary with system parameters.
+        do_yield: Whether to yield the simulation data as a tuple (sim_idx, sample).
+        save_raw_data: Whether to save the raw data (as images and labels) to the dataset_dir.
     """
     # initiate ODE term from `ode_fn`
     ode_term = ODETerm(ode_fn)
 
     # initiate time steps array of samples
-    ts = jnp.arange(0, horizon + dt, dt)
+    ts = jnp.arange(0, horizon_dim * dt, step=dt)
 
     # if the simulation time-step is not given, initialize it to the same value as the sample time step
     if sim_dt is None:
@@ -60,7 +64,7 @@ def collect_dataset(
 
     # dataset directory
     dataset_path = Path(dataset_dir)
-    if dataset_path.exists():
+    if dataset_path.exists() and save_raw_data:
         # empty the directory
         shutil.rmtree(dataset_dir)
     # (re)create the directory
@@ -111,26 +115,41 @@ def collect_dataset(
             # dimension of configuration space
             n_q = n_x // 2
 
-            # folder to save the simulation data
-            sim_dir = dataset_path / f"sim-{sim_idx}"
-            sim_dir.mkdir(parents=True, exist_ok=True)
-
+            # define labels dict
             labels = dict(t_ts=ts, x_ts=x_ts)
-            # save the labels in dataset_dir
-            jnp.savez(file=str(sim_dir / "labels.npz"), **labels)
 
+            if save_raw_data:
+                # folder to save the simulation data
+                sim_dir = dataset_path / f"sim-{sim_idx}"
+                sim_dir.mkdir(parents=True, exist_ok=True)
+
+                # save the labels in dataset_dir
+                jnp.savez(file=str(sim_dir / "labels.npz"), **labels)
+
+            rendering_ts = []
             for time_idx in range(x_ts.shape[0]):
                 # configuration for current time step
                 q = x_ts[time_idx, :n_q]
 
                 # render the image
                 img = rendering_fn(q)
+                rendering_ts.append(img)
 
-                # save the image
-                cv2.imwrite(str(sim_dir / f"rendering_time_idx-{time_idx}.jpeg"), img)
+                if save_raw_data:
+                    # save the image
+                    cv2.imwrite(str(sim_dir / f"rendering_time_idx-{time_idx}.jpeg"), img)
 
                 # update sample index
                 sample_idx += 1
+
+            # merge labels with image and id
+            sample = labels | {
+                "id": sim_idx,
+                "rendering_ts": rendering_ts,
+            }
+
+            if do_yield:
+                yield sim_idx, sample
 
             # update progress bar
             bar()
